feat(e2e-test) move e2e scripts into scripts/e2e
