+++ packages/core/scripts/e2e/README.md
