+++ packages/core/scripts/e2e/e2e-test-page.js
