+++ packages/core/scripts/e2e/snapshot.js
