--- packages/core/scripts/run-e2e.js
@@ -1,3 +1,0 @@
-'use strict';
-
-require('./e2e').run();
